In That we apply CNN for finding the Accuracy
Pomegranates are also a fruit that grows with a high yield 
in several Asian countries and is one amongst the foremost profit-gaining fruits within 
the market. However, due to vary conditions, the plants area unit infected 
by varied diseases that destroy the whole crop yield very. So, the work proposes an 
image method and neural network ways to agitate the foremost issues with 
physiopathology, i.e., detection and classification of eudemonia. The pomegranate 
fruit is additionally attributed to the very fact that its leaves area unit plagued 
by diseases caused by plants and weather. These diseases area unit like blight 
microorganisms, plant spots, seed rot, and leaf spot. The system uses 
some photos for employment, some for testing functions, and so on. The 
final accuracy of this method is ninetieth. The results area unit proven to be correct 
and satisfactory in distinction to manual grading. Hopefully, they'll take a strong rise 
in establishing themselves among the market united of the foremost economic 
processes. To use a deep learning technique that might kind the fruits 
into traditional and abnormal supported the options like fruit color, variety of fruit 
spots, and form of the fruit. A combined CNN deep learning model is applied to 
classify a collection of 6519 fruits into 2 categories, particularly traditional and 
abnormal. Convolution Neural Networks (CNN) area unit applied for deep feature 
extraction, that is employed to sight the category supported extracted options. The 
sensitivity of malady detection was lower with the lesser accessibility of increased 
detection ways for detective work malady in earlier stages. The problem with these 
varied existing algorithms is that their accuracy has been reduced since some sources 
of approaches, and functionalities which will facilitate to resolve analytic and 
prophetic analysis accurately.
Keywords: Deep Learning, CNN, Classification, Microorganisms, Accuracy, 
Detection